refactor: address thermo-nuclear review for playable client
- Unify fetchGameConfig client; delete duplicate game-config.ts
- Add useGameSession hook with WS log sync (action_applied + schedule)
- Extract lobbyResponses.ts; route all lobby JSON through buildLobbyResponse
- Collapse board UI duplication; optimize BoardGrid indexes and diagonal view
- Add typed gameLogDisplay formatter; shared workerD1Stub for integration tests
