+++ packages/web/src/components/BoardGrid.tsx
@@ -1,41 +1,55 @@
 import type { GameState } from "@oligopoly/validation";
 import { tileLabel } from "../lib/boardDisplay";
-import { currentActorId } from "../lib/gameUi";
 
 type BoardGridProps = {
   state: GameState;
   tileNames: Map<string, string>;
   myPlayerId: string | null;
+  actorId: string | null;
 };
 
 const BOTTOM_ROW = Array.from({ length: 11 }, (_, index) => index);
 const RIGHT_COLUMN = Array.from({ length: 10 }, (_, index) => 11 + index);
 const TOP_ROW = Array.from({ length: 11 }, (_, index) => 31 - index);
 const LEFT_COLUMN = Array.from({ length: 8 }, (_, index) => 32 + index);
+const DIAGONAL_POSITIONS = ["D1", "D2", "D3", "D4", "D5"];
 
-function tileOwnerId(state: GameState, position: number): string | null {
-  const tile = state.tiles?.find((entry) => entry.position === position);
-  return tile?.ownerId ?? null;
+function indexPlayersByPosition(state: GameState) {
+  const map = new Map<string, NonNullable<GameState["players"]>>();
+  for (const player of state.players ?? []) {
+    const key = String(player.position);
+    const group = map.get(key) ?? [];
+    group.push(player);
+    map.set(key, group);
+  }
+  return map;
 }
 
-function playersAtPosition(state: GameState, position: number | string) {
-  return state.players?.filter((player) => player.position === position) ?? [];
+function indexTileOwners(state: GameState) {
+  const map = new Map<number, string>();
+  for (const tile of state.tiles ?? []) {
+    if (typeof tile.position === "number" && tile.ownerId) {
+      map.set(tile.position, tile.ownerId);
+    }
+  }
+  return map;
 }
 
 function BoardCell({
   position,
-  state,
+  ownerId,
+  occupants,
+  actorId,
   tileNames,
   myPlayerId,
 }: {
   position: number;
-  state: GameState;
+  ownerId: string | null;
+  occupants: NonNullable<GameState["players"]>;
+  actorId: string | null;
   tileNames: Map<string, string>;
   myPlayerId: string | null;
 }) {
-  const ownerId = tileOwnerId(state, position);
-  const occupants = playersAtPosition(state, position);
-  const actorId = currentActorId(state);
   const isMine = ownerId !== null && ownerId === myPlayerId;
   const isActorHere = occupants.some((player) => player.playerId === actorId);
 
@@ -68,7 +82,18 @@ function BoardCell({
   );
 }
 
-export function BoardGrid({ state, tileNames, myPlayerId }: BoardGridProps) {
+export function BoardGrid({
+  state,
+  tileNames,
+  myPlayerId,
+  actorId,
+}: BoardGridProps) {
+  const occupantsByPosition = indexPlayersByPosition(state);
+  const ownerByPosition = indexTileOwners(state);
+  const diagonalOccupants = DIAGONAL_POSITIONS.flatMap(
+    (position) => occupantsByPosition.get(position) ?? [],
+  );
+
   return (
     <div className="boardGrid">
       <div className="boardGridRow boardGridRowTop">
@@ -76,7 +101,9 @@ export function BoardGrid({ state, tileNames, myPlayerId }: BoardGridProps) {
           <BoardCell
             key={position}
             position={position}
-            state={state}
+            ownerId={ownerByPosition.get(position) ?? null}
+            occupants={occupantsByPosition.get(String(position)) ?? []}
+            actorId={actorId}
             tileNames={tileNames}
             myPlayerId={myPlayerId}
           />
@@ -89,7 +116,9 @@ export function BoardGrid({ state, tileNames, myPlayerId }: BoardGridProps) {
             <BoardCell
               key={position}
               position={position}
-              state={state}
+              ownerId={ownerByPosition.get(position) ?? null}
+              occupants={occupantsByPosition.get(String(position)) ?? []}
+              actorId={actorId}
               tileNames={tileNames}
               myPlayerId={myPlayerId}
             />
@@ -97,14 +126,21 @@ export function BoardGrid({ state, tileNames, myPlayerId }: BoardGridProps) {
         </div>
 
         <div className="boardGridCenter">
-          <p className="muted">Round {state.round}</p>
-          <p>
-            <strong>{state.phase ?? "unknown"}</strong>
-          </p>
-          <p className="muted">
-            Free market: {state.settings?.currencySymbol ?? "¤"}
-            {state.freeMarketPool ?? 0}
-          </p>
+          <p className="muted">Diagonal path</p>
+          {diagonalOccupants.length > 0 ? (
+            <ul className="boardGridDiagonalList">
+              {diagonalOccupants.map((player) => (
+                <li key={player.playerId}>
+                  {player.playerId === myPlayerId
+                    ? "You"
+                    : (player.displayName ?? player.playerId.slice(0, 8))}{" "}
+                  at {tileLabel(player.position, tileNames)}
+                </li>
+              ))}
+            </ul>
+          ) : (
+            <p className="muted">No players on diagonal</p>
+          )}
         </div>
 
         <div className="boardGridColumn">
@@ -112,7 +148,9 @@ export function BoardGrid({ state, tileNames, myPlayerId }: BoardGridProps) {
             <BoardCell
               key={position}
               position={position}
-              state={state}
+              ownerId={ownerByPosition.get(position) ?? null}
+              occupants={occupantsByPosition.get(String(position)) ?? []}
+              actorId={actorId}
               tileNames={tileNames}
               myPlayerId={myPlayerId}
             />
@@ -125,7 +163,9 @@ export function BoardGrid({ state, tileNames, myPlayerId }: BoardGridProps) {
           <BoardCell
             key={position}
             position={position}
-            state={state}
+            ownerId={ownerByPosition.get(position) ?? null}
+            occupants={occupantsByPosition.get(String(position)) ?? []}
+            actorId={actorId}
             tileNames={tileNames}
             myPlayerId={myPlayerId}
           />
