+++ packages/web/src/hooks/useGameRealtime.ts
@@ -1,4 +1,4 @@
-import type { GameState } from "@oligopoly/validation";
+import type { GameLogEntry, GameState } from "@oligopoly/validation";
 import {
   GameRealtimeEventSchema,
   GameStateSchema,
@@ -6,8 +6,14 @@ import {
 import { useEffect, useRef, useState } from "react";
 import { gameWebSocketUrl } from "../api/games";
 
+export type GameSessionUpdate = {
+  state: GameState;
+  logEntries?: GameLogEntry[];
+  source: string;
+};
+
 type UseGameRealtimeOptions = {
-  onState?: (state: GameState, source: string) => void;
+  onUpdate?: (update: GameSessionUpdate) => void;
 };
 
 export function useGameRealtime(
@@ -16,9 +22,8 @@ export function useGameRealtime(
 ) {
   const [wsStatus, setWsStatus] = useState("disconnected");
   const [turnDeadline, setTurnDeadline] = useState<number | null>(null);
-  const [lastEvent, setLastEvent] = useState<string | null>(null);
-  const onStateRef = useRef(options.onState);
-  onStateRef.current = options.onState;
+  const onUpdateRef = useRef(options.onUpdate);
+  onUpdateRef.current = options.onUpdate;
 
   useEffect(() => {
     if (!gameId) {
@@ -37,31 +42,41 @@ export function useGameRealtime(
         const parsed = GameRealtimeEventSchema.safeParse(
           JSON.parse(String(event.data)),
         );
-        if (parsed.success) {
-          const message = parsed.data;
-          if (message.type === "game.action_applied" && "state" in message) {
-            const nextState = GameStateSchema.parse(message.state);
-            onStateRef.current?.(nextState, "Realtime state update");
-            return;
-          }
-          if (message.type === "game.snapshot" && "payload" in message) {
-            const nextState = GameStateSchema.parse(message.payload);
-            onStateRef.current?.(nextState, "Realtime snapshot");
-            return;
-          }
-          if (message.type === "game.timer" && "deadlineAt" in message) {
-            setTurnDeadline(message.deadlineAt ?? null);
-            return;
-          }
+        if (!parsed.success) return;
+
+        const message = parsed.data;
+        if (message.type === "game.action_applied" && "state" in message) {
+          onUpdateRef.current?.({
+            state: GameStateSchema.parse(message.state),
+            logEntries: message.logEntries,
+            source: "Realtime state update",
+          });
+          return;
+        }
+        if (message.type === "game.schedule" && "state" in message) {
+          onUpdateRef.current?.({
+            state: GameStateSchema.parse(message.state),
+            source: "Realtime schedule update",
+          });
+          return;
+        }
+        if (message.type === "game.snapshot" && "payload" in message) {
+          onUpdateRef.current?.({
+            state: GameStateSchema.parse(message.payload),
+            source: "Realtime snapshot",
+          });
+          return;
+        }
+        if (message.type === "game.timer" && "deadlineAt" in message) {
+          setTurnDeadline(message.deadlineAt ?? null);
         }
       } catch {
-        // Fall through to raw event logging.
+        // Ignore malformed websocket payloads.
       }
-      setLastEvent(`Realtime event: ${event.data}`);
     };
 
     return () => socket.close();
   }, [gameId]);
 
-  return { wsStatus, turnDeadline, lastEvent, setLastEvent };
+  return { wsStatus, turnDeadline };
 }
